requirements.txt
Django==4.1.3
djangorestframework==3.14.0
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
sqlparse==0.4.3
